# chore: simplify computation of pow for each sumcheck round