We used to construct a separate data structure to store the "constant contribution" of the pow polynomials : at round l this will be  the product of the partial evaluation result `c_l` (i.e. `pow(u_0,...,u_{l-1})`where `u_0,...,u_{l-1}`) are the verifier challenges and the elements of `pow(\vec{beta})`  not containing `\beta_0,...,\beta_l`. This can be simplified, additionally avoiding unnecessary field multiplications. This does not affect benchmarks:
### BEFORE
```
--------------------------------------------------------------------------------
Benchmark                        Time                   CPU   
--------------------------------------------------------------------------------
ClientIVCBench/Full/6      21590 ms        16489 ms  
```
### AFTER
``` 
--------------------------------------------------------------------------------
Benchmark                      Time             CPU   
--------------------------------------------------------------------------------
ClientIVCBench/Full/6      21571 ms    16454 ms 
```